refactor: Organised the Navbar Section
