+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Layout from "./pages/Page";
+import Layout from "./pages/Layout";
 import Home from "./pages/Home";
 import Upload from "./pages/Studio";
 import Video from "./pages/Video";
