+++ frontend/src/components/Navbar.jsx
@@ -21,6 +21,9 @@ import LoginComponent from "./LoginComponent";
 import SignupComponent from "./SignupComponent";
 import { useNavigate } from "react-router-dom";
 import { useAuth } from "../Authentication/AuthContext";
+import SearchBar from "./SearchBar";
+import ProfileMenu from "./ProfileMenu";
+import AuthModal from "./AuthModal";
 
 export default function Navbar({ onMenuClick }) {
   const { profilePic, logout, isAuthenticated } = useAuth();
@@ -71,23 +74,7 @@ export default function Navbar({ onMenuClick }) {
             </g>
           </svg>
         </a>
-        <form className="flex flex-grow py-1 max-w-xl mx-auto">
-          <input
-            className="form-input px-4 py-1 w-full border border-gray-300 rounded-l-full"
-            id="search_key"
-            name="search_key"
-            type="search"
-            placeholder="Search"
-            aria-label="Search"
-          />
-          <button
-            className="bg-gray-200 px-6 py-1 rounded-r-full"
-            title="Search"
-            type="submit"
-          >
-            <FontAwesomeIcon icon={faMagnifyingGlass} />
-          </button>
-        </form>
+        <SearchBar/>
         <div className="flex items-center space-x-7 px-4">
           {isAuthenticated ? (
             <a href="/studio" className="text-gray-700" title="Upload">
@@ -107,78 +94,7 @@ export default function Navbar({ onMenuClick }) {
             <FontAwesomeIcon icon={faBell} />
           </button>
           {isAuthenticated ? (
-            <Popover>
-              <PopoverTrigger>
-                <button
-                  className="text-gray-700 rounded-full w-6 h-6"
-                  type="button"
-                  title="Profile"
-                >
-                  <img
-                    src={profilePic}
-                    alt="Profile"
-                    className="rounded-full w-6 h-6"
-                  />
-                </button>
-              </PopoverTrigger>
-              <PopoverContent>
-                <div className="px-2 py-2 w-[240px]">
-                  <a
-                    className="relative flex items-center w-full h-12 px-3 mt-2 rounded hover:bg-gray-300"
-                    onClick={logout}
-                  >
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      height="24px"
-                      viewBox="0 -960 960 960"
-                      width="24px"
-                      fill="#1f1f1f"
-                    >
-                      <path d="M200-120q-33 0-56.5-23.5T120-200v-560q0-33 23.5-56.5T200-840h280v80H200v560h280v80H200Zm440-160-55-58 102-102H360v-80h327L585-622l55-58 200 200-200 200Z" />
-                    </svg>
-                    <span className="ml-2 text-sm font-medium">LogOut</span>
-                  </a>
-                  <a
-                    className="flex items-center w-full h-12 px-3 mt-2 rounded hover:bg-gray-300"
-                    href="/createChannel"
-                  >
-                    <svg
-                      version="1.1"
-                      id="Layer_1"
-                      x="0px"
-                      y="0px"
-                      width="20"
-                      height="20"
-                      viewBox="0 0 122.88 107.78"
-                      style={{ enableBackground: "new 0 0 122.88 107.78" }}
-                    >
-                      <g>
-                        <path d="M2.1,81.26c-1.16,0-2.1-0.94-2.1-2.1c0-1.16,0.94-2.1,2.1-2.1h9.09V12.44H2.1c-1.16,0-2.1-0.94-2.1-2.1 c0-1.16,0.94-2.1,2.1-2.1h57.12V2.1c0-1.16,0.94-2.1,2.1-2.1s2.1,0.94,2.1,2.1v6.14h57.12c1.16,0,2.1,0.94,2.1,2.1 c0,1.16-0.94,2.1-2.1,2.1h-8.95v64.62h9.18c1.16,0,2.1,0.94,2.1,2.1c0,1.16-0.94,2.1-2.1,2.1H63.75v9.6 c0.1,0.07,0.2,0.15,0.29,0.23l13.12,12.48c0.84,0.8,0.87,2.12,0.07,2.96c-0.8,0.84-2.12,0.87-2.96,0.07L63.75,96.6v9.08 c0,1.16-0.94,2.1-2.1,2.1c-1.16,0-2.1-0.94-2.1-2.1V96.2L48.6,106.61c-0.84,0.8-2.17,0.76-2.96-0.07c-0.8-0.84-0.77-2.17,0.07-2.96 L58.84,91.1c0.21-0.2,0.46-0.35,0.71-0.45v-9.39H2.1L2.1,81.26z M61.44,28.04c9.28,0,16.81,7.53,16.81,16.81 c0,9.28-7.53,16.81-16.81,16.81c-9.28,0-16.81-7.53-16.81-16.81C44.63,35.57,52.16,28.04,61.44,28.04L61.44,28.04z M70.05,44.94 l-13.4-8.72v17.89L70.05,44.94L70.05,44.94z M15.39,77.06h92V12.44h-92V77.06L15.39,77.06z" />
-                      </g>
-                    </svg>
-                    <span className="ml-2 text-sm font-medium">
-                      Your Channel
-                    </span>
-                  </a>
-                  <a
-                    className="relative flex items-center w-full h-12 px-3 mt-2 rounded hover:bg-gray-300"
-                    href="#"
-                  >
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      x="0px"
-                      y="0px"
-                      width="20"
-                      height="20"
-                      viewBox="0 0 50 50"
-                    >
-                      <path d="M 22.205078 2 A 1.0001 1.0001 0 0 0 21.21875 2.8378906 L 20.246094 8.7929688 C 19.076509 9.1331971 17.961243 9.5922728 16.910156 10.164062 L 11.996094 6.6542969 A 1.0001 1.0001 0 0 0 10.708984 6.7597656 L 6.8183594 10.646484 A 1.0001 1.0001 0 0 0 6.7070312 11.927734 L 10.164062 16.873047 C 9.583454 17.930271 9.1142098 19.051824 8.765625 20.232422 L 2.8359375 21.21875 A 1.0001 1.0001 0 0 0 2.0019531 22.205078 L 2.0019531 27.705078 A 1.0001 1.0001 0 0 0 2.8261719 28.691406 L 8.7597656 29.742188 C 9.1064607 30.920739 9.5727226 32.043065 10.154297 33.101562 L 6.6542969 37.998047 A 1.0001 1.0001 0 0 0 6.7597656 39.285156 L 10.648438 43.175781 A 1.0001 1.0001 0 0 0 11.927734 43.289062 L 16.882812 39.820312 C 17.936999 40.39548 19.054994 40.857928 20.228516 41.201172 L 21.21875 47.164062 A 1.0001 1.0001 0 0 0 22.205078 48 L 27.705078 48 A 1.0001 1.0001 0 0 0 28.691406 47.173828 L 29.751953 41.1875 C 30.920633 40.838997 32.033372 40.369697 33.082031 39.791016 L 38.070312 43.291016 A 1.0001 1.0001 0 0 0 39.351562 43.179688 L 43.240234 39.287109 A 1.0001 1.0001 0 0 0 43.34375 37.996094 L 39.787109 33.058594 C 40.355783 32.014958 40.813915 30.908875 41.154297 29.748047 L 47.171875 28.693359 A 1.0001 1.0001 0 0 0 47.998047 27.707031 L 47.998047 22.207031 A 1.0001 1.0001 0 0 0 47.160156 21.220703 L 41.152344 20.238281 C 40.80968 19.078827 40.350281 17.974723 39.78125 16.931641 L 43.289062 11.933594 A 1.0001 1.0001 0 0 0 43.177734 10.652344 L 39.287109 6.7636719 A 1.0001 1.0001 0 0 0 37.996094 6.6601562 L 33.072266 10.201172 C 32.023186 9.6248101 30.909713 9.1579916 29.738281 8.8125 L 28.691406 2.828125 A 1.0001 1.0001 0 0 0 27.705078 2 L 22.205078 2 z M 23.056641 4 L 26.865234 4 L 27.861328 9.6855469 A 1.0001 1.0001 0 0 0 28.603516 10.484375 C 30.066026 10.848832 31.439607 11.426549 32.693359 12.185547 A 1.0001 1.0001 0 0 0 33.794922 12.142578 L 38.474609 8.7792969 L 41.167969 11.472656 L 37.835938 16.220703 A 1.0001 1.0001 0 0 0 37.796875 17.310547 C 38.548366 18.561471 39.118333 19.926379 39.482422 21.380859 A 1.0001 1.0001 0 0 0 40.291016 22.125 L 45.998047 23.058594 L 45.998047 26.867188 L 40.279297 27.871094 A 1.0001 1.0001 0 0 0 39.482422 28.617188 C 39.122545 30.069817 38.552234 31.434687 37.800781 32.685547 A 1.0001 1.0001 0 0 0 37.845703 33.785156 L 41.224609 38.474609 L 38.53125 41.169922 L 33.791016 37.84375 A 1.0001 1.0001 0 0 0 32.697266 37.808594 C 31.44975 38.567585 30.074755 39.148028 28.617188 39.517578 A 1.0001 1.0001 0 0 0 27.876953 40.3125 L 26.867188 46 L 23.052734 46 L 22.111328 40.337891 A 1.0001 1.0001 0 0 0 21.365234 39.53125 C 19.90185 39.170557 18.522094 38.59371 17.259766 37.835938 A 1.0001 1.0001 0 0 0 16.171875 37.875 L 11.46875 41.169922 L 8.7734375 38.470703 L 12.097656 33.824219 A 1.0001 1.0001 0 0 0 12.138672 32.724609 C 11.372652 31.458855 10.793319 30.079213 10.427734 28.609375 A 1.0001 1.0001 0 0 0 9.6328125 27.867188 L 4.0019531 26.867188 L 4.0019531 23.052734 L 9.6289062 22.117188 A 1.0001 1.0001 0 0 0 10.435547 21.373047 C 10.804273 19.898143 11.383325 18.518729 12.146484 17.255859 A 1.0001 1.0001 0 0 0 12.111328 16.164062 L 8.8261719 11.46875 L 11.523438 8.7734375 L 16.185547 12.105469 A 1.0001 1.0001 0 0 0 17.28125 12.148438 C 18.536908 11.394293 19.919867 10.822081 21.384766 10.462891 A 1.0001 1.0001 0 0 0 22.132812 9.6523438 L 23.056641 4 z M 25 17 C 20.593567 17 17 20.593567 17 25 C 17 29.406433 20.593567 33 25 33 C 29.406433 33 33 29.406433 33 25 C 33 20.593567 29.406433 17 25 17 z M 25 19 C 28.325553 19 31 21.674447 31 25 C 31 28.325553 28.325553 31 25 31 C 21.674447 31 19 28.325553 19 25 C 19 21.674447 21.674447 19 25 19 z"></path>
-                    </svg>
-                    <span className="ml-2 text-sm font-medium">Settings</span>
-                  </a>
-                </div>
-              </PopoverContent>
-            </Popover>
+            <ProfileMenu profilePic={profilePic} logout={logout} />
           ) : (
             <button
               className="text-gray-700 rounded-full w-6 h-6"
@@ -190,43 +106,7 @@ export default function Navbar({ onMenuClick }) {
             </button>
           )}
 
-          <Modal isOpen={isOpen} backdrop={"blur"} onOpenChange={onOpenChange}>
-            <ModalContent>
-              {(onClose) => (
-                <>
-                  {showLogin ? (
-                    <>
-                      <LoginComponent onClose={onClose} />
-                      <p className="mt-4 text-center text-sm pb-6">
-                        Don't have an account?{" "}
-                        <Link
-                          onClick={() => setShowLogin(!showLogin)}
-                          className="text-blue-600 hover:underline"
-                          style={{ cursor: "pointer" }}
-                        >
-                          Create account
-                        </Link>
-                      </p>
-                    </>
-                  ) : (
-                    <>
-                      <SignupComponent setShowLogin={setShowLogin} />
-                      <p className="mt-4 text-center text-sm pb-6">
-                        Already have an account?{" "}
-                        <Link
-                          onClick={() => setShowLogin(!showLogin)}
-                          className="text-blue-600 hover:underline"
-                          style={{ cursor: "pointer" }}
-                        >
-                          Login
-                        </Link>
-                      </p>
-                    </>
-                  )}
-                </>
-              )}
-            </ModalContent>
-          </Modal>
+          <AuthModal onOpenChange={onOpenChange} isOpen={isOpen}/>
         </div>
       </div>
     </nav>
